Specify dependencies explicitly
Relying on Minecraft dependencies is _probabaly_ OK, but doesn't match
my understanding of best practices, which is to declare what is
necessary so a transitive change doesn't break the project.

Given the Fabric ecosystem, it is assumed that Minecraft, Fabric API,
and Mixins will be available.  Everything else must be declared.