Write Instructions to a file
This will be the basis of pre-computed, or "baked", animations.  They
will be loaded from resource files using an `Identifier` schema yet to
be determined (likely "apel:$modId:$animationName").

Right now, all lines are rendered in absolute terms, which is not useful
for pre-computed animations in most expected cases.  Animations are
likely to be in a position relative to the player or some other entity
in the world.

The baking renderer adds one method beyond the ApelServerRenderer: a
`finish()` method so that the `dataOutputStream` can be closed properly.
There is probably a better way to do this that's interface-compatible;
perhaps, a default method, but that is a last resort.