Enhance builder on ParticleModel
Clean up some of the internals.  There's very little chance the
model needs to change at runtime; if that's the case, a new
ParticleModel should be created and swapped in to whatever animator is
being used.  Also, pass the full ObjectModel to the interceptor, since
it _could_ have polylines to change, too.