Scale the ParticleCone
Cone scaling requires an update to its basic rendering due to needing
the cone's base, currently a circle, to be treated as an ellipse.
Therefore, add a `stretch` parameter so `ParticleCone` can render
similarly to `ParticleCircle` and allow an elliptical base, which is
possible when a cone is scaled differently in `x` and `z` directions.